refactor(providers): break k8s->aws import via neutral NativeSpec protocol
K8sNativeSpecService imported from orb.providers.aws, coupling the two
providers. Extract the shared contract to a provider-neutral location so
each provider depends on the abstraction, not on aws.

- Add orb/providers/base/native_spec_protocol.py with NativeSpecServiceProtocol
- aws_native_spec_service: remove inline Protocol definition, import from base,
  re-export for backward compatibility
- k8s_native_spec_service: import NativeSpecServiceProtocol from providers.base
  instead of providers.aws
- aws_handler_factory: update TYPE_CHECKING import to providers.base

Both cross-provider grep checks now return empty.
